ergpy/requirements.txt
JPype1>=1.3.0
requests>=2.27.1
stubgenj>=0.2.5